Use "singleTask" launch mode in the demo activity
Without "singleTask", after the user makes a color selection, `DemoActivity.onCreate()` is called with a `null` value for `savedInstanceState`. It's not possible to restore the state of the fragment.

With "singleTask", after the user makes a color selection, `DemoActivity` is brought to the front without going through `onCreate()`: it's the same instance of the activity. The fragment doesn't lose it's state. When the fragment resumes, it knows it's resuming from a browser switch, and proceeds to process the deep link.