Use the browser switch with custom tabs in the same activity task as `BraintreeFragment`'s activity
Motivation:

When 3DS is done in an external browser inside another activity stack, sometimes the users try to navigate back to the app using back or recents:
* Either they navigate back to the app instead of browser after reading their 3DS code (in the sms app or bank app).
* Or, perhaps more rarely, the automatic deep link redirection after a successful code entry doesn't work, and users try to manually navigate back to the app using back or recents, instead of tapping the big blue "RETURN TO APP" button.

These scenarios result in the purchase being canceled.

This PR uses custom tabs in the same activity stack as the app. When the 3DS input succeeds, and the automatic redirection works, the custom tabs are automatically closed, and the calling activity (containing `BrowserSwitchFragment`) is brought to the top of the activity stack.

If the automatic redirection doesn't succeed, we still have the possible problem of a confused user who has to tap on the "RETURN TO APP" button. But since we're in the same activity stack as the app, the user can't navigate away, so the purchase can't be canceled.

Technical details:

* Don't add these flags when launching the browser intent: `FLAG_ACTIVITY_MULTIPLE_TASK`, `FLAG_ACTIVITY_CLEAR_TASK`, `FLAG_ACTIVITY_NEW_TASK`
* Launch the browser from the activity context, not from the application context (we can't use the application context anymore as that requires `FLAG_ACTIVITY_NEW_TASK`)
* Tell `BrowserSwitchActivity` which intent to launch when it finishes (`DemoActivity` in the demo example). This is done in two places:
  - The "normal" scenario is from `BrowserSwitchClient`, just before it launches custom tabs.
  - The other scenario is an edge case, in case the Android OS is aggressively killing our app process (reproducible by toggling "Don't keep activities" and "No background processes" in system developer settings). In this case, the application class tells `BrowserSwitchActivity` which intent to launch.

Note: I tried a few other approaches. For example, using an activity lifecycle callback listener to detect when `BrowserSwitchActivity` finishes, to relaunch `BrowserSwitchFragment`'s activity. These had various issues which can be found in the commit comments of the initial PR #18.